Renamed Hal to HAL to comply with general naming scheme and to make it compilable on clean Linux
